\name{BayesPLS-package}
\alias{BayesPLS-package}
\docType{package}
\title{
\packageTitle{BayesPLS}
}
\description{
\packageDescription{BayesPLS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesPLS}
\packageIndices{BayesPLS}
To be filled in.
}
\author{
\packageAuthor{BayesPLS}

Maintainer: \packageMaintainer{BayesPLS}
}
\references{
Helland et al. 2012. SJS. Details to be filled in.
}

\keyword{ package}

\examples{
## See ?BayesPLS
}
