\name{predict.BayesPLS}
\alias{predict.BayesPLS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction function for Bayes PLS
}
\description{
Simple linear predictor for new observations.
}
\usage{
predict.BayesPLS(obj, newX)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{
  The fitted Bayes PLS object.
  }
  \item{newX}{
  A matrix containing new predictor observations.
  }
}


\author{
Solve Sæbø
}

\examples{
## See ?BayesPLS
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }% use one of  RShowDoc("KEYWORDS")
\keyword{ regression }% __ONLY ONE__ keyword per line
